src/hawkbit-client: Fix free storage space calculation overflowing
The number of available bytes of storage calculated can overflow when there is too much free space.
On a machine where long is only 32 bits in size the overflow occurs with more than about 2 GB available.
The get_available_space() function then reports a negative value and the rollout fails with "Not enough free space".

Signed-off-by: Florian Schrank <f.schrank@osb-ag.de>